# contracts/labby-openapi.yaml
openapi: 3.1.0
info:
  title: Labby API
  version: 0.1.0
paths:
...
  /api/v2/labby/channels/:
    get:
      tags:
      - sales-channels
      summary: List Channels
      operationId: list_channels_api_v2_labby_channels__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelsResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - sales-channels
      summary: Create Channel
      operationId: create_channel_api_v2_labby_channels__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesChannelCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v2/labby/sales/channels/:
    get:
      tags:
      - sales-channels
      summary: List Channels
      operationId: list_channels_api_v2_labby_sales_channels__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelsResponse'
      security:
      - HTTPBearer: []
    post:
      tags:
      - sales-channels
      summary: Create Channel
      operationId: create_channel_api_v2_labby_sales_channels__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesChannelCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v2/labby/channels/{channel_id}:
    get:
      tags:
      - sales-channels
      summary: Get Channel
      operationId: get_channel_api_v2_labby_channels__channel_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - sales-channels
      summary: Update Channel
      operationId: update_channel_api_v2_labby_channels__channel_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesChannelUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sales-channels
      summary: Delete Channel
      operationId: delete_channel_api_v2_labby_channels__channel_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/channels/{channel_id}:
    get:
      tags:
      - sales-channels
      summary: Get Channel
      operationId: get_channel_api_v2_labby_sales_channels__channel_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - sales-channels
      summary: Update Channel
      operationId: update_channel_api_v2_labby_sales_channels__channel_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesChannelUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sales-channels
      summary: Delete Channel
      operationId: delete_channel_api_v2_labby_sales_channels__channel_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/channels/{channel_id}/status:
    get:
      tags:
      - sales-channels
      summary: Channel Status
      operationId: channel_status_api_v2_labby_channels__channel_id__status_get
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/channels/{channel_id}/status:
    get:
      tags:
      - sales-channels
      summary: Channel Status
      operationId: channel_status_api_v2_labby_sales_channels__channel_id__status_get
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/channels/{channel_id}/connect:
    post:
      tags:
      - sales-channels
      summary: Connect Channel
      operationId: connect_channel_api_v2_labby_channels__channel_id__connect_post
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/SalesChannelConnectRequest'
              - type: 'null'
              title: Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/channels/{channel_id}/connect:
    post:
      tags:
      - sales-channels
      summary: Connect Channel
      operationId: connect_channel_api_v2_labby_sales_channels__channel_id__connect_post
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/SalesChannelConnectRequest'
              - type: 'null'
              title: Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/channels/{channel_id}/disconnect:
    post:
      tags:
      - sales-channels
      summary: Disconnect Channel
      operationId: disconnect_channel_api_v2_labby_channels__channel_id__disconnect_post
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/channels/{channel_id}/disconnect:
    post:
      tags:
      - sales-channels
      summary: Disconnect Channel
      operationId: disconnect_channel_api_v2_labby_sales_channels__channel_id__disconnect_post
      security:
      - HTTPBearer: []
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesChannelConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SalesChannelConnectRequest:
      properties:
        bot_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Bot Token
        phone_number_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number Id
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
        waba_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Waba Id
        guild_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Guild Id
        greeting:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Greeting
        position:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Position
        widget_color:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Widget Color
      type: object
      title: SalesChannelConnectRequest
    SalesChannelConnectionResponse:
      properties:
        status:
          type: string
          enum:
          - desconectado
          - conectando
          - conectado
          - erro
          title: Status
        message:
          anyOf:
          - type: string
          - type: 'null'
          title: Message
        qr_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Qr Code
        instance_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Instance Name
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        numero:
          anyOf:
          - type: string
          - type: 'null'
          title: Numero
        bot_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Bot Username
        bot_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Bot Name
        oauth_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Oauth Url
        widget_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Widget Id
        snippet:
          anyOf:
          - type: string
          - type: 'null'
          title: Snippet
      type: object
      required:
      - status
      title: SalesChannelConnectionResponse
    SalesChannelCreateRequest:
      properties:
        tipo:
          type: string
          enum:
          - whatsapp_evolution
          - whatsapp_cloud
          - telegram
          - discord
          - web_chatbot
          title: Tipo
        nome:
          type: string
          maxLength: 120
          minLength: 1
          title: Nome
        config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config
      type: object
      required:
      - tipo
      - nome
      title: SalesChannelCreateRequest
    SalesChannelDeleteResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        message:
          type: string
          title: Message
      type: object
      required:
      - id
      - message
      title: SalesChannelDeleteResponse
    SalesChannelResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        tipo:
          type: string
          enum:
          - whatsapp_evolution
          - whatsapp_cloud
          - telegram
          - discord
          - web_chatbot
          title: Tipo
        nome:
          type: string
          title: Nome
        status:
          type: string
          enum:
          - desconectado
          - conectando
          - conectado
          - erro
          title: Status
        config:
          additionalProperties: true
          type: object
          title: Config
        webhook_configured:
          type: boolean
          title: Webhook Configured
        ultimo_evento_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ultimo Evento At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - tipo
      - nome
      - status
      - config
      - webhook_configured
      title: SalesChannelResponse
    SalesChannelStatusResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tipo:
          type: string
          enum:
          - whatsapp_evolution
          - whatsapp_cloud
          - telegram
          - discord
          - web_chatbot
          title: Tipo
        nome:
          type: string
          title: Nome
        status:
          type: string
          enum:
          - desconectado
          - conectando
          - conectado
          - erro
          title: Status
        numero:
          anyOf:
          - type: string
          - type: 'null'
          title: Numero
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        bot_username:
          anyOf:
          - type: string
          - type: 'null'
          title: Bot Username
        guild_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Guild Name
        widget_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Widget Id
        config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config
        ultimo_evento_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ultimo Evento At
      type: object
      required:
      - id
      - tipo
      - nome
      - status
      title: SalesChannelStatusResponse
    SalesChannelUpdateRequest:
      properties:
        nome:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Nome
        config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Config
      type: object
      title: SalesChannelUpdateRequest
    SalesChannelsResponse:
      properties:
        channels:
          items:
            $ref: '#/components/schemas/SalesChannelResponse'
          type: array
          title: Channels
      type: object
      required:
      - channels
      title: SalesChannelsResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer